requirements.txt
# requeriments.txt
fastapi
uvicorn
motor
pydantic
openpyxl
python-multipart
python-dotenv
python-jose[cryptography]
passlib[bcrypt]

